+++ src/logic/Entitys/Perspectiva.java
@@ -62,7 +62,7 @@ public class Perspectiva {
     public float[] calculate_MDr_IMD(){
         float[] md = new float[2];
 
-        md[0] = cant_ptos / cant_preguntas;
+        md[0] = (float) cant_ptos / (float)cant_preguntas;
         md[1] = (md[0] / 4) * 100;
 
         return md;
